% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_clean}
\alias{data_clean}
\title{Cleaned Nordic plant trait dataset}
\format{
A data frame with the following variables:
\describe{
\item{species}{Species name}
\item{seed_mass}{Seed mass}
\item{SLA}{Specific leaf area}
\item{leaf_CN_ratio}{Leaf carbon-to-nitrogen ratio}
\item{leaf_N}{Leaf nitrogen}
\item{height}{Plant height}
\item{leaf_area}{Leaf area}
\item{rooting_depth}{Rooting depth}
\item{LDMC}{Leaf dry matter content}
\item{woodiness}{Woodiness category}
}
}
\source{
NordicTraits dataset
}
\usage{
data_clean
}
\description{
A cleaned dataset of plant traits used by the Shiny app.
}
\keyword{datasets}
