% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_trait_plot}
\alias{make_trait_plot}
\title{Build the interactive ggiraph scatterplot for the chosen plant trait data.}
\usage{
make_trait_plot(df, input)
}
\arguments{
\item{df}{filtered data frame used for plotting.}

\item{input}{A shiny input object or input-like list wiht
\code{x_trait}, \code{y_trait}, \code{log_xy}, and \code{trend_line}.}
}
\value{
A ggiraph widget.
}
\description{
Build the interactive ggiraph scatterplot for the chosen plant trait data.
}
\keyword{internal}
