% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{NordicPlantTraitExplorer-package}
\alias{NordicPlantTraitExplorer}
\alias{NordicPlantTraitExplorer-package}
\title{NordicPlantTraitExplorer package}
\description{
R package that launches an interactive shiny app for exploring relationships among plant functional traits. The package includes a dataset of plant trait values for over 3000 species from the Nordic region. Can visualise trait relationships to explore ecological strategies such as leaf economics and rescouse useage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NoraFanebust/NordicPlantTraitExplorer}
  \item Report bugs at \url{https://github.com/NoraFanebust/NordicPlantTraitExplorer/issues}
}

}
\author{
\strong{Maintainer}: Nora Fanebust \email{nora.fanebust@hotmail.com}

}
\keyword{internal}
