% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Build the Shiny app server}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{A Shiny input object.}

\item{output}{A Shiny output object.}

\item{session}{A Shiny session object.}
}
\value{
No return value. Called for its side effects in a Shiny app.
}
\description{
Build the Shiny app server
}
\keyword{internal}
