% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{trait_choices}
\alias{trait_choices}
\title{Chosen traits for the app
Returns the column names that can be selected for the x- and y-axes.}
\usage{
trait_choices()
}
\value{
A character vector of trait variable names.
}
\description{
Chosen traits for the app
Returns the column names that can be selected for the x- and y-axes.
}
\keyword{internal}
