% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\title{Creates the user interface for the plant trait explorer shiny app.}
\usage{
ui()
}
\value{
A Shiny UI definition.
}
\description{
Creates the user interface for the plant trait explorer shiny app.
}
\keyword{internal}
